[refactor] 서비스에서 동적으로 LIMIT, type 결정하도록 수정 (#114)
