+++ src/main/java/konkuk/thip/recentSearch/application/service/manager/RecentSearchCreateManager.java
@@ -1,5 +1,6 @@
 package konkuk.thip.recentSearch.application.service.manager;
 
+import konkuk.thip.recentSearch.adapter.out.jpa.RecentSearchType;
 import konkuk.thip.recentSearch.application.port.out.RecentSearchCommandPort;
 import konkuk.thip.recentSearch.application.port.out.RecentSearchQueryPort;
 import konkuk.thip.recentSearch.domain.RecentSearch;
@@ -12,15 +13,13 @@ import java.time.LocalDateTime;
 @RequiredArgsConstructor
 public class RecentSearchCreateManager {
 
-    private static final String USER_SEARCH_TERM = "사용자 검색";
-
     private final RecentSearchCommandPort recentSearchCommandPort;
     private final RecentSearchQueryPort recentSearchQueryPort;
 
-    public void saveRecentSearchByUser(Long userId, String keyword) {
+    public void saveRecentSearchByUser(Long userId, String keyword, RecentSearchType type) {
 
         // 동일 조건 (userId + keyword + type) 검색 기록이 이미 있는지 확인
-        recentSearchQueryPort.findRecentSearchByKeywordAndUserId(keyword, userId)
+        recentSearchQueryPort.findRecentSearchByKeywordAndUserId(keyword, userId, type)
                 .ifPresentOrElse(
                         existingRecentSearch -> {
                             // 이미 존재하면 createdAt만 갱신
@@ -29,7 +28,7 @@ public class RecentSearchCreateManager {
                         },
                         () -> {
                             // 없으면 새로 저장
-                            RecentSearch userRecentSearch = RecentSearch.withoutId(keyword, USER_SEARCH_TERM, userId);
+                            RecentSearch userRecentSearch = RecentSearch.withoutId(keyword, type, userId);
                             recentSearchCommandPort.save(userRecentSearch);
                         }
                 );
